Move all temporal related operations into client_operations
